Reorganized the GTK3 theming code.
I moved the Stylus code (and its subsequently compiled CSS file) to
the new 'gtk-3.0' folder in order to make this a standalone theme
that does not require the old CSS Blues theme in order to function.
The system wants this particular organization hierarchy in order to
automatically parse 3rd party themes, so I don't really have any
choice but to follow the layout seen here.

Many of the raw resources (i.e. PNG images) that are present in this
commit exist in order to preserve the theme's original functionality.
However, as these resources do not belong to me (they were originally
part of the Ambiance-Blackout theme from www.ravefinity.com), they
will be removed and/or replaced over time with my own custom elements.

I also modified the build script to work with these changes and to
produce a file called 'gtk.css' instead of 'gtk-theme.css', which is
what the OS expects to find when hunting for theme entry points.