Partially fixed the transparency issue with CSD windows.
This required manually patching, building, and installing Compiz so
that CSD windows don't show up with a black backing window. I suspect
that those black windows might be frame windows that Compiz reparents
others into, but I'm not totally sure as of now.

Although the newly patched Compiz works pretty well, it has the odd
side-effect of changing the selector that ultimately gets applied to
CSD windows in Compiz. Instead of the '.solid-csd' that was present
before (beginning with Ubuntu 15.10; never saw it in 15.04), these
windows are again tagged with the simple '.csd' selector (as they
were in 15.04). Presumably, this is how things are supposed to work;
I know Ubuntu's handling of CSD windows attempts to override that
designation and make them server-side decorated, but I don't know
where exactly that occurs. Without caring enough to scour Ubuntu's
and Compiz's codebase, I'm just going to guess that Compiz plays a
role there.

To partially address this issue, I've divided CSD styling into two
separate selectors: one for CSD windows recognized by the patched
Compiz ('.csd') and one for unrecognized windows ('.solid-csd').
It's not perfect, but it's the best I can do for now.