Fixed a bug introduced by removing the 'button-text-color' function.
I removed that function from the style library because for a while
it looked like only buttons were using them. Now something here wants
it as well, so I've copied that function to the top of this file.

I also moved the '.nautilus-canvas-item' styling code over to the
'Minutiae' file since it's specific to Nautilus.