Removed the importing of old control assets.
This wasn't working anyway and was generating minor errors in the
console, so I just commented it out for the time being. At some
point soon I'll go through and create my own icons for the UI
widgets that are currently being affected (pretty much only radio
and checkbox controls).