Major cleanup of styling code and additional fixes.
I almost completely rewrote this file today, and got rid of a bunch
of stuff that wasn't really necessary. I also fixed a long-standing
issue with some menus and text (e.g. in several places in Evolution)
where submenus and buttons that open menus caused all text to behave
like a mouse was hovering over it.