Major updates for compatibility with GTK+ v3.22.
Around v3.20 of GTK+, the CSS styling system for UI widgets was
completely overhauled, rendering this code obsolete. This update fixes a
majority of the issues and restores basic functionality to the theme.
However, some work remains to be done before all of t he styling
features can be brought back.